Refactor: Move common Manim utilities to manim_common.py
